+++ lib/src/loading_state_handler_widget.dart
@@ -3,101 +3,100 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 
 import 'current_state_enum.dart';
+import 'typedefs.dart';
 
 /// A widget that handles loading, error, and empty states.
 ///
 class LoadingStateHandlerWidget extends StatefulWidget {
   /// Default loading widget builder.
   ///
-  /// This callback will be called when [loading] is true, and the default
+  /// This callback will be called when [loading] is true or [currentState] is [CurrentStateEnum.loading], and the default
   /// widget is displayed.
   ///
   /// The default implementation is to display a [CircularProgressIndicator] in
   /// the center of the screen.
   ///
   /// If not null, the default loading widget will be displayed when [loading] is
-  /// true.
+  /// true or [currentState] is [CurrentStateEnum.loading].
   ///
   /// The default value is null, which means the default loading widget will not
   /// be displayed.
-  static Widget Function(BuildContext, String?)? _defaultLoadingBuilder;
+  static DefaultLoadingBuilder _defaultLoadingBuilder;
 
   /// Default error widget builder.
   ///
-  /// This callback will be called when [error] is true, and the default
+  /// This callback will be called when [error] is true or [currentState] is [CurrentStateEnum.error], and the default
   /// widget is displayed.
   ///
   /// The default implementation is to display a [Text] widget in the center of
   /// the screen with the error message.
   ///
   /// If not null, the default error widget will be displayed when [error] is
-  /// true.
+  /// true or [currentState] is [CurrentStateEnum.error].
   ///
   /// The default value is null, which means the default error widget will not
   /// be displayed.
-  static Widget Function(
-          BuildContext, String?, Widget, Duration, VoidCallback?)?
-      _defaultErrorBuilder;
+  static DefaultErrorBuilder _defaultErrorBuilder;
 
   /// Default empty widget builder.
   ///
-  /// This callback will be called when [empty] is true, and the default
+  /// This callback will be called when [empty] is true or [currentState] is [CurrentStateEnum.empty], and the default
   /// widget is displayed.
   ///
   /// The default implementation is to display a [Text] widget in the center of
   /// the screen with the empty message.
   ///
   /// If not null, the default empty widget will be displayed when [empty] is
-  /// true.
+  /// true or [currentState] is [CurrentStateEnum.empty].
   ///
   /// The default value is null, which means the default empty widget will not
   /// be displayed.
-  static Widget Function(BuildContext, String?)? _defaultEmptyBuilder;
+  static DefaultEmptyBuilder _defaultEmptyBuilder;
 
   /// Default on loading callback.
   ///
-  /// This callback will be called when [loading] is true.
+  /// This callback will be called when [loading] is true or [currentState] is [CurrentStateEnum.loading].
   ///
   /// If not null, the default on loading callback will be called when [loading]
-  /// is true.
+  /// is true or [currentState] is [CurrentStateEnum.loading].
   ///
   /// The default value is null, which means the default on loading callback will
   /// not be called.
-  static void Function(BuildContext, String?)? _defaultOnLoading;
+  static DefaultOnLoading _defaultOnLoading;
 
   /// Default on error callback.
   ///
-  /// This callback will be called when [error] is true.
+  /// This callback will be called when [error] is true or [currentState] is [CurrentStateEnum.error].
   ///
   /// If not null, the default on error callback will be called when [error] is
-  /// true.
+  /// true or [currentState] is [CurrentStateEnum.error].
   ///
   /// The default value is null, which means the default on error callback will
   /// not be called.
-  static void Function(BuildContext, String?, VoidCallback?)? _defaultOnError;
+  static DefaultOnError _defaultOnError;
 
   /// Default on empty callback.
   ///
-  /// This callback will be called when [empty] is true.
+  /// This callback will be called when [empty] is true or [currentState] is [CurrentStateEnum.empty].
   ///
   /// If not null, the default on empty callback will be called when [empty] is
-  /// true.
+  /// true or [currentState] is [CurrentStateEnum.empty].
   ///
   /// The default value is null, which means the default on empty callback will
   /// not be called.
-  static void Function(BuildContext, String?)? _defaultOnEmpty;
+  static DefaultOnEmpty _defaultOnEmpty;
 
   /// Default on data callback.
   ///
   /// This callback will be called when [loading] is false, [error] is false,
-  /// and [empty] is false.
+  /// and [empty] is false or [currentState] is [CurrentStateEnum.data].
   ///
   /// If not null, the default on data callback will be called when [loading] is
-  /// false, [error] is false, and [empty] is false.
+  /// false, [error] is false, and [empty] is false or [currentState] is [CurrentStateEnum.data].
   ///
   /// The default value is null, which means the default on data callback will
   /// not be called.
-  static void Function(BuildContext, String?)? _defaultOnData;
+  static DefaultOnData _defaultOnData;
 
   /// To disable widget changes globally.
   ///
@@ -218,7 +217,7 @@ class LoadingStateHandlerWidget extends StatefulWidget {
 
   /// The widget to display when the state is loading.
   ///
-  /// If not null, the [loadingWidget] will be displayed when [loading] is true.
+  /// If not null, the [loadingWidget] will be displayed when [loading] is true or [currentState] is [CurrentStateEnum.loading].
   ///
   /// The default value is null, which means the default loading widget will be
   /// used.
@@ -226,7 +225,7 @@ class LoadingStateHandlerWidget extends StatefulWidget {
 
   /// The widget to display when the state is error.
   ///
-  /// If not null, the [errorWidget] will be displayed when [error] is true.
+  /// If not null, the [errorWidget] will be displayed when [error] is true or [currentState] is [CurrentStateEnum.error].
   ///
   /// The default value is null, which means the default error widget will be
   /// used.
@@ -234,7 +233,7 @@ class LoadingStateHandlerWidget extends StatefulWidget {
 
   /// The widget to display when the state is empty.
   ///
-  /// If not null, the [emptyWidget] will be displayed when [empty] is true.
+  /// If not null, the [emptyWidget] will be displayed when [empty] is true or [currentState] is [CurrentStateEnum.empty].
   ///
   /// The default value is null, which means the default empty widget will be
   /// used.
@@ -261,7 +260,7 @@ class LoadingStateHandlerWidget extends StatefulWidget {
   /// The widget to display when the state is normal.
   ///
   /// This widget will be displayed when [loading] is false, [error] is false,
-  /// and [empty] is false.
+  /// and [empty] is false or [currentState] is [CurrentStateEnum.normal].
   final Widget child;
 
   /// A callback function that will be called when the state is error.
@@ -277,12 +276,7 @@ class LoadingStateHandlerWidget extends StatefulWidget {
   ///
   /// The default value is null, which means the default error widget will be
   /// used.
-  final Function(
-    Function(BuildContext, String?, VoidCallback?)?,
-    BuildContext,
-    String?,
-    VoidCallback?,
-  )? onError;
+  final OnError onError;
 
   /// A callback function that will be called when the state is empty.
   ///
@@ -297,11 +291,7 @@ class LoadingStateHandlerWidget extends StatefulWidget {
   ///
   /// The default value is null, which means the default empty widget will be
   /// used.
-  final Function(
-    Function(BuildContext, String?)?,
-    BuildContext,
-    String?,
-  )? onEmpty;
+  final OnEmpty onEmpty;
 
   /// A callback function that will be called when the state is loading.
   ///
@@ -316,11 +306,7 @@ class LoadingStateHandlerWidget extends StatefulWidget {
   ///
   /// The default value is null, which means the default loading widget will be
   /// used.
-  final Function(
-    Function(BuildContext, String?)?,
-    BuildContext,
-    String?,
-  )? onLoading;
+  final OnLoading onLoading;
 
   /// A callback function that will be called when the state is normal.
   ///
@@ -335,11 +321,7 @@ class LoadingStateHandlerWidget extends StatefulWidget {
   ///
   /// The default value is null, which means the default data widget will be
   /// used.
-  final Function(
-    Function(BuildContext, String?)?,
-    BuildContext,
-    String?,
-  )? onData;
+  final OnData onData;
 
   /// The callback to be executed when retry is attempted.
   final VoidCallback? onRetry;
@@ -422,14 +404,13 @@ class LoadingStateHandlerWidget extends StatefulWidget {
     TextStyle? defaultRetryBtnTextStyle,
     TextStyle? defaultRetryMessageStyle,
     ButtonStyle? defaultRetryButtonStyle,
-    Widget Function(BuildContext, String?)? defaultLoadingBuilder,
-    Widget Function(BuildContext, String?, Widget, Duration, VoidCallback?)?
-        defaultErrorBuilder,
-    Widget Function(BuildContext, String?)? defaultEmptyBuilder,
-    Function(BuildContext, String?, VoidCallback?)? defaultOnError,
-    Function(BuildContext, String?)? defaultOnEmpty,
-    Function(BuildContext, String?)? defaultOnLoading,
-    Function(BuildContext, String?)? defaultOnData,
+    DefaultLoadingBuilder defaultLoadingBuilder,
+    DefaultErrorBuilder defaultErrorBuilder,
+    DefaultEmptyBuilder defaultEmptyBuilder,
+    DefaultOnError defaultOnError,
+    DefaultOnEmpty defaultOnEmpty,
+    DefaultOnLoading defaultOnLoading,
+    DefaultOnData defaultOnData,
   }) {
     _defaultRetryCooldown = defaultRetryCooldown ?? _defaultRetryCooldown;
 
@@ -478,15 +459,15 @@ class _LoadingStateHandlerWidgetState extends State<LoadingStateHandlerWidget> {
   /// [LoadingStateHandlerWidget] and returns the appropriate widget based on
   /// the state.
   ///
-  /// If [loading] is true, the method returns the [loadingWidget] if it is not
+  /// If [loading] is true or [currentState] is [CurrentStateEnum.loading], the method returns the [loadingWidget] if it is not
   /// null, otherwise it returns the default loading widget which is a
   /// [Center] widget with a [CircularProgressIndicator].
   ///
-  /// If [error] is true, the method returns the [errorWidget] if it is not null,
+  /// If [error] is true or [currentState] is [CurrentStateEnum.error], the method returns the [errorWidget] if it is not null,
   /// otherwise it returns the default error widget which is a [Center] widget
   /// with the text 'Error'.
   ///
-  /// If [empty] is true, the method returns the [emptyWidget] if it is not null,
+  /// If [empty] is true or [currentState] is [CurrentStateEnum.empty], the method returns the [emptyWidget] if it is not null,
   /// otherwise it returns the default empty widget which is a [Center] widget
   /// with the text 'Empty'.
   ///
@@ -633,10 +614,10 @@ class _LoadingStateHandlerWidgetState extends State<LoadingStateHandlerWidget> {
   ///
   /// The methods are called in the following order:
   ///
-  /// 1. [onLoading] if [loading] is true.
-  /// 2. [onError] if [error] is true.
-  /// 3. [onEmpty] if [empty] is true.
-  /// 4. [onData] if [loading], [error], and [empty] are all false.
+  /// 1. [onLoading] if [loading] is true or [currentState] is [CurrentStateEnum.loading].
+  /// 2. [onError] if [error] is true or [currentState] is [CurrentStateEnum.error].
+  /// 3. [onEmpty] if [empty] is true or [currentState] is [CurrentStateEnum.empty].
+  /// 4. [onData] if [loading], [error], and [empty] are all false or [currentState] is [CurrentStateEnum.data].
   ///
   /// If the methods are not provided, the default methods are called.
   /// If the default methods are not provided, the methods are not called.
@@ -677,6 +658,30 @@ class _LoadingStateHandlerWidgetState extends State<LoadingStateHandlerWidget> {
   }
 
   /// Builds the retry button.
+  ///
+  /// This method returns a [Column] containing:
+  ///
+  /// 1. If [_remainingCooldown] is greater than 0, a [Text] with the
+  ///    retry message. The retry message is optionally provided by the
+  ///    user through the [retryMessage] constructor parameter.
+  ///    If [retryMessage] is null, the default retry message is used.
+  ///
+  /// 2. An [ElevatedButton] that when pressed, calls the [onRetry] callback
+  ///    and starts the retry cooldown timer.
+  ///
+  /// The [ElevatedButton] is only enabled when [_remainingCooldown]
+  /// is 0, i.e., the cooldown has expired.
+  ///
+  /// The button is styled according to the [retryButtonStyle] constructor
+  /// parameter. If [retryButtonStyle] is null, the default button style is
+  /// used.
+  ///
+  /// The button's text is styled according to the [retryBtnTextStyle]
+  /// constructor parameter. If [retryBtnTextStyle] is null, the default
+  /// button text style is used.
+  ///
+  /// The button's text is the [retryBtnText] constructor parameter. If
+  /// [retryBtnText] is null, the default button text is used.
   Widget _buildRetryButton() {
     if (!widget.enableRetry || widget.onRetry == null) {
       return const SizedBox.shrink();
@@ -684,6 +689,7 @@ class _LoadingStateHandlerWidgetState extends State<LoadingStateHandlerWidget> {
 
     return Column(
       mainAxisSize: MainAxisSize.min,
+      spacing: 8,
       children: [
         if (_remainingCooldown > 0)
           Text(
@@ -716,22 +722,42 @@ class _LoadingStateHandlerWidgetState extends State<LoadingStateHandlerWidget> {
   }
 
   /// Starts the retry cooldown timer.
+  ///
+  /// This method is called when the retry button is pressed. It sets the
+  /// [_remainingCooldown] to the value of [retryCooldown] constructor
+  /// parameter, or to the default value if [retryCooldown] is null.
+  ///
+  /// Then, it starts a timer that decrements [_remainingCooldown] every
+  /// second. When [_remainingCooldown] reaches 0, the timer is cancelled.
+  ///
+  /// The timer is stored in [_retryCooldownTimer] to be cancelled if the
+  /// retry button is pressed again before the cooldown has expired.
   void _startRetryCooldown() {
     if (!widget.enableRetry || widget.onRetry == null) return;
 
+    /// if retryCooldown is null, use the default value
     _remainingCooldown = (widget.retryCooldown ??
             LoadingStateHandlerWidget._defaultRetryCooldown)
         .inSeconds;
 
+    /// cancel the previous timer if it exists
     _retryCooldownTimer?.cancel();
-    _retryCooldownTimer = Timer.periodic(const Duration(seconds: 1), (timer) {
-      setState(() {
-        if (_remainingCooldown > 0) {
-          _remainingCooldown--;
-        } else {
-          timer.cancel();
-        }
-      });
-    });
+
+    /// start the new timer
+    _retryCooldownTimer = Timer.periodic(
+      const Duration(seconds: 1),
+      (timer) {
+        setState(
+          () {
+            if (_remainingCooldown > 0) {
+              _remainingCooldown--;
+            } else {
+              /// when the cooldown has expired, cancel the timer
+              timer.cancel();
+            }
+          },
+        );
+      },
+    );
   }
 }